+++ src/Lesson_8/Worker.java
@@ -2,7 +2,7 @@ package Lesson_8;
 
 public class Worker extends Employee{
     @Override
-    public int getSalary(){
+    public double getSalary(){
         return getBaseSalary();
     }
 }
